# Demo - Added a new MainScene that asynchronously switches from a MenuScene, to the GamePlayScene - Added a new MenuScene - Added a GameLoadSpinner that shows while scenes are being swapped on a background thread
# Engine

Update SceneHardChange to load scenes on a background thread
- updated BaseScene to handle destruction and post instantiation behaviors

Cleaned up uses of context
- `context` is no longer a requirement of base scenes

Removed uses of parent
- child classes no longer take a parent in their constructor for registration purposes. Instead, a parent must register the children directly.

Fixed an issue where a SceneManager would run logic and drawing for BaseScenes that had those abilities disabled.

Updated the SceneManager SubScene to not add scenes if their id already exists in the registry.

# Other
- Added assets for demo loading screen and menu